requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.60.0
requests>=2.25.0
beautifulsoup4>=4.9.0
sentence-transformers>=2.2.0
transformers>=4.31.0
torch>=1.10.0
scipy>=1.7.0
huggingface-hub>=0.16.0 
chromadb==1.5.1
ollama==0.6.1
python-dotenv==1.0.1